[refactor] FrameSet 과 관련된 도메인을 set 패키지로 이동
