+++ src/main/java/bowling/domain/FrameResults.java
@@ -1,5 +1,7 @@
 package bowling.domain;
 
+import bowling.domain.set.FrameSet;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
