+++ src/main/java/bowling/domain/set/FrameSet.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.set;
 
 import bowling.domain.state.State;
